Fix C23 compatibility and 64-bit type issues
- Replace custom Boolean typedef with standard _Bool and <stdbool.h> in box.h/box.c
- Add function prototypes for messageToUserWithObj, messageToUser (error.h)
- Add function prototypes for _copyPacket, dumpPacket (packet.h)
- Add function prototype for biOp with proper types (biop.h)
- Fix function pointer types in libstg: int (*) -> long (*) for tagName2ID functions
- Fix STGFindAssert calls to use tagName2ID() conversion for attrName parameter
  * STGAssert->name stores token IDs as tagged pointers, not strings
  * Added (char*)(long)tagName2ID() casts in stgcall.c (3 places)
- Update compareDocViewShell to use _Bool with proper prototype (vw.h/vw.c)

All changes fix -Wdeprecated-non-prototype warnings and ensure 64-bit compatibility.
Builds successfully with 'make check-64bit' without errors.

+++ src/libStyle/libstg.h
@@ -67,8 +67,8 @@ STGAssert* assertAttr(STGAssert** firstAssert, char* name, char* val);
 struct ParseContext;
 int nextToken(struct ParseContext* ctx, char** s);
 
-STGLib* STG_init(int (*tagNameCmp_f)(), int (*tagName2ID_f)(), char* (*tagID2Name_f)(), 
-                 int (*tagAttrNameCmp_f)(), int (*tagAttrName2ID_f)(), char* (*tagAttrID2Name_f)());
+STGLib* STG_init(int (*tagNameCmp_f)(), long (*tagName2ID_f)(), char* (*tagID2Name_f)(), 
+                 int (*tagAttrNameCmp_f)(), long (*tagAttrName2ID_f)(), char* (*tagAttrID2Name_f)());
 STGGroup* STG_makeGroup(STGLib* lib, char* gspec);
 int STG_findStyle(STGGroup* group, char* context[], int contextCount, STGResult results[], int maxResults);
 STGAssert* STGFindAssert(STGMajor* major, char* attrName);
